apps/voice-agent/requirements.txt
# ── Voice pipeline ────────────────────────────────────────────────────────────
pipecat-ai[deepgram,openai,anthropic,elevenlabs,silero]>=0.0.50

# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.1

# ── Database ──────────────────────────────────────────────────────────────────
motor>=3.6.0
pymongo>=4.10.0

# ── Cache / pub-sub ───────────────────────────────────────────────────────────
redis[asyncio]>=5.2.0

# ── HTTP client ───────────────────────────────────────────────────────────────
httpx>=0.27.0

# ── Config ────────────────────────────────────────────────────────────────────
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# ── Logging ───────────────────────────────────────────────────────────────────
structlog>=24.4.0

# ── Storage ───────────────────────────────────────────────────────────────────
aioboto3>=13.0.0

# ── LLM SDKs (direct, in addition to pipecat wrappers) ───────────────────────
openai>=1.54.0
anthropic>=0.40.0

# ── Audio ─────────────────────────────────────────────────────────────────────
# audioop is included in Python 3.12 stdlib — no extra package needed
numpy>=1.26.0
